feat: adds working collapsable search bar
This is how I really like the search bar, as it resembles other apps. But, I will check other ways of doing this search bar. As sometimes less is more.

Also, this search bar is a bit complex, and maybe, just maybe, a bit unnecessary. Still, it is a good exercise, and I finally did it!