wip: invoice drafting and loading
I'm not smart enough for this, I fear. I feel I'm going in the right track-ish. Maybe... Idk. I feel sometimes it is too much, that I don't know enough. That some of this might be very wrong. I'm tired, sick, and a bit sad. Maybe tomorrow will be a better day.